# docs/openapi.yaml
...
openapi: 3.0.3
info:
  title: PGFPlotsGenerator API
  description: '智能图表生成系统后端接口。除标注外均需 `Authorization: Bearer <token>`。'
  version: "1.0"
  contact:
    name: PGFPlotsGenerator
servers:
  - url: http://localhost:3000
    description: 本地默认后端地址（见 src/config.js）
...
security:
  - BearerAuth: []
paths:
...
  /api/datasets:
    get:
      tags: [数据集]
      summary: 我的数据集列表（关键词搜索）
      parameters:
        - name: keyword
          in: query
          schema: { type: string }
      responses:
        "200":
          description: 成功（code 风格）
          content:
            application/json:
              schema:
                type: object
                properties:
                  code: { type: integer, example: 200 }
                  message: { type: string }
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/DataFile" }
    post:
      tags: [数据集]
      summary: 上传数据集（multipart/form-data，≤100MB）
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [name, description, file]
              properties:
                name: { type: string, example: 全国人口数据 }
                description: { type: string }
                file: { type: string, format: binary }
      responses:
        "200":
          description: 上传成功（code 风格）
          content:
            application/json:
              schema:
                type: object
                properties:
                  code: { type: integer, example: 200 }
                  message: { type: string }
                  data: { $ref: "#/components/schemas/DataFile" }
        "400":
          description: 名称/描述为空或未选择文件
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CodeFail" }
  /api/datasets/{id}/update:
    post:
      tags: [数据集]
      summary: 修改数据集名称/描述
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                description: { type: string }
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  code: { type: integer, example: 200 }
                  message: { type: string }
                  data: { type: object }
        "400":
          description: 未提供任何可更新字段
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CodeFail" }
  /api/datasets/{id}:
    delete:
      tags: [数据集]
      summary: 删除数据集（级联删除文件）
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: 删除成功
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CodeSuccess" }
  /api/datasets/download/{id}:
    get:
      tags: [数据集]
      summary: 下载数据集原文件
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: 文件下载
          content:
            application/octet-stream: {}

  # ===================== 对话 /api/conversations =====================
  /api/conversations:
    get:
      tags: [对话]
      summary: 会话列表（含消息数与最后一条预览）
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/ConversationItem" }
    post:
      tags: [对话]
      summary: 新建会话
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                title: { type: string, default: 新对话, maxLength: 255 }
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data:
                    type: object
                    properties:
                      conversation_id: { type: integer }
                      title: { type: string }
  /api/conversations/{id}:
    put:
      tags: [对话]
      summary: 重命名会话
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title]
              properties:
                title: { type: string, maxLength: 255 }
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MessageSuccess" }
    delete:
      tags: [对话]
      summary: 删除会话（事务先删消息再删会话）
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MessageSuccess" }
        "404":
          description: 会话不存在
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MessageFail" }
  /api/conversations/{id}/messages:
    get:
      tags: [对话]
      summary: 获取会话消息（按时间正序）
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/ConversationMessage" }

  # ===================== 反馈 /api/feedback =====================
  /api/feedback:
    post:
      tags: [反馈]
      summary: 用户提交反馈
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [type, content]
              properties:
                type:
                  type: string
                  enum: [suggestion, ui, bug, other]
                  description: 建议/界面/缺陷/其他
                content: { type: string, maxLength: 100 }
      responses:
        "201":
          description: 提交成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  message: { type: string }
                  data:
                    type: object
                    properties:
                      feedback_id: { type: integer }
        "400":
          description: 类型/内容为空、超长或类型非法
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MessageFail" }
    get:
      tags: [反馈]
      summary: 管理员获取全部反馈（筛选 type/日期，分页）
      description: 需管理员权限（服务端 checkAdmin 校验 role=admin）。
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 10 }
        - name: type
          in: query
          schema: { type: string, enum: [suggestion, ui, bug, other] }
        - name: startDate
          in: query
          schema: { type: string, format: date }
        - name: endDate
          in: query
          schema: { type: string, format: date }
      responses:
        "200":
          description: 成功（含用户名字段）
          content:
            application/json:
              schema: { $ref: "#/components/schemas/FeedbackListResponse" }
  /api/feedback/user/my-feedbacks:
    get:
      tags: [反馈]
      summary: 用户查看自己的反馈（含已回复/待回复状态）
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 10 }
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema: { $ref: "#/components/schemas/FeedbackListResponse" }
  /api/feedback/{id}:
    get:
      tags: [反馈]
      summary: 管理员查看反馈详情
      description: 需管理员权限。
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  message: { type: string }
                  data: { $ref: "#/components/schemas/FeedbackItem" }
        "404":
          description: 反馈不存在
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MessageFail" }
    delete:
      tags: [反馈]
      summary: 管理员删除反馈
      description: 需管理员权限。
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: 删除成功
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MessageSuccess" }
  /api/feedback/{id}/reply:
    put:
      tags: [反馈]
      summary: 管理员回复反馈
      description: 回复后自动向反馈者写入一条定向 notice（feedback_id 唯一索引幂等）。需管理员权限。
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [answer]
              properties:
                answer: { type: string, maxLength: 500 }
      responses:
        "200":
          description: 回复成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  message: { type: string }
                  data: { $ref: "#/components/schemas/FeedbackItem" }
        "400":
          description: 回复为空或超 500 字
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MessageFail" }
        "404":
          description: 反馈不存在
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MessageFail" }

  # ===================== 通知 /api/notice =====================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: '用户/管理员登录返回的 JWT（请求头 `Authorization: Bearer <token>`）'
  schemas:
...
    MessageSuccess:
      type: object
      properties:
        success: { type: boolean, example: true }
        message: { type: string }
    MessageFail:
      type: object
      properties:
        success: { type: boolean, example: false }
        message: { type: string }
    CodeSuccess:
      type: object
      properties:
        code: { type: integer, example: 200 }
        message: { type: string }
        data: { nullable: true }
    CodeFail:
      type: object
      properties:
        code: { type: integer, example: 400 }
        message: { type: string }
...
    DataFile:
      type: object
      description: 数据集条目（列表/上传返回字段）
      properties:
        data_id: { type: integer }
        user_id: { type: integer }
        data_name: { type: string }
        data_size: { type: integer }
        size: { type: string, description: 格式化后的可读大小 }
        description: { type: string, nullable: true }
        upload_time: { type: string }
        update_time: { type: string }
        file_name: { type: string }
        file_path: { type: string }
        mimetype: { type: string }
    ConversationItem:
      type: object
      properties:
        conversation_id: { type: integer }
        title: { type: string }
        message_count: { type: integer }
        last_message: { type: string }
        created_at: { type: string }
        updated_at: { type: string }
    ConversationMessage:
      type: object
      properties:
        message_id: { type: integer }
        role: { type: string, enum: [user, assistant] }
        content: { type: string, nullable: true }
        chart_code: { type: string, nullable: true }
        history_id: { type: integer, nullable: true }
        selected_files: { type: string, nullable: true, description: 数据集快照 JSON 字符串 }
        created_at: { type: string }
    FeedbackItem:
      type: object
      description: 反馈条目（管理员列表额外含 username）
      properties:
        feedback_id: { type: integer }
        user_id: { type: integer }
        username: { type: string, description: 管理员视角返回 }
        type: { type: string, enum: [suggestion, ui, bug, other] }
        content: { type: string }
        answer: { type: string, nullable: true }
        status: { type: string, description: 用户视角：已回复/待回复 }
        feedback_time: { type: string }
        answer_time: { type: string, nullable: true }
    FeedbackListResponse:
      type: object
      properties:
        success: { type: boolean, example: true }
        message: { type: string }
        data:
          type: array
          items: { $ref: "#/components/schemas/FeedbackItem" }
        pagination:
          type: object
          properties:
            page: { type: integer }
            limit: { type: integer }
            total: { type: integer }
            pages: { type: integer }
...